Move the code of Environment to Engine
