+++ tests/bootstrap.php
@@ -3,4 +3,3 @@
 include __DIR__ . '/../autoload.php';
 
 \Minz\Configuration::load('test', __DIR__ . '/test_app');
-\Minz\Environment::initialize();
